Minor modification to support Bash 4
It seems that `$(sed 's| |,|g' <<< $BOXES)` trims the `$BOXES` variable before `sed` acts upon it, but this is not the case in Bash 4, so the script breaks, as osascript complains due to the extra comma in the list of boxes. This fixes the problem for all versions of Bash.